feat(status-bar): goal celebration, UV thresholds, and threshold-UI polish
- Goal celebration rework: steps/sleep/distance count UPWARD toward the
  pair — Close (warn slot, green outline #55FF00) then Goal reached
  (danger slot, green fill). below_is_worse is retired to false
  everywhere (C + JS contract `goal` flag); seeds reordered upward and
  chip labels read Close/Goal.
- UV thresholds (kind 7): settings blob 27 -> 29 bytes (exact-length
  validation keeps stale blobs inert), STATUS_LEVELS_UINT8 widened to
  2 LE wire bytes with UV at bits 8-9, slider 0-12 with seeds 6/8.
  The persist accessors carry the widened word as int — a uint8_t
  setter truncated the high byte and silently dropped UV highlights
  (regression-pinned in the contract test). Weather bundle 526 B
  (10 B headroom), Clay 391 B.
- Config-UI: the slot edit pencil is now an outlined Goal/Warn text
  button showing the ring+dot pair when enabled; wind/gust/AQI hints no
  longer recite units/scales.

+++ src/c/appendix/persist.c
@@ -490,7 +490,7 @@ int persist_get_status_levels(void) {
     return persist_read_int(STATUS_LEVELS);
 }
 
-bool persist_set_status_levels(uint8_t levels) {
+bool persist_set_status_levels(int levels) {
     return write_int_if_changed(STATUS_LEVELS, levels);
 }
